requirements.txt
PyYAML==3.11
argparse==1.2.1
ascii-graph==1.0.1
beautifulsoup4==4.4.1
coverage==4.0.1
decorator==4.0.4
funcsigs==0.4
matplotlib==1.4.3
mock==1.3.0
networkx==1.10
nose==1.3.7
numpy==1.10.1
pbr==1.8.1
py==1.4.30
pyparsing==2.0.3
pytest==2.8.2
pytest-cov==2.2.0
python-dateutil==2.4.2
pytz==2015.6
requests==2.8.1
selenium==2.48.0
six==1.10.0
vk==1.5.3
wsgiref==0.1.2
